+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend_user_service_registration_functions.go
@@ -2,9 +2,11 @@ package docker
 
 import (
 	"context"
+	"fmt"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/user_service_registration"
 	"github.com/kurtosis-tech/stacktrace"
+	"time"
 )
 
 func (backend *DockerKurtosisBackend) CreateUserServiceRegistration(
@@ -15,7 +17,7 @@ func (backend *DockerKurtosisBackend) CreateUserServiceRegistration(
 	backend.serviceRegistrationMutex.Lock()
 	defer backend.serviceRegistrationMutex.Unlock()
 
-	registrationsForEnclave, found := backend.serviceRegistrations[enclaveId]
+	registrationsForEnclave, found := backend.serviceRegistrationsIndex[enclaveId]
 	if !found {
 		return nil, stacktrace.NewError(
 			"Service registration information isn't being tracked for enclave '%v', which indicates that no free IP address provider was " +
@@ -54,24 +56,70 @@ func (backend *DockerKurtosisBackend) CreateUserServiceRegistration(
 		}
 	}()
 
-	registration := user_service_registration.NewUserServiceRegistration(enclaveId, serviceId, ipAddr)
+	// TODO Switch this, and all other GUIDs, to a UUID!
+	guid := user_service_registration.UserServiceRegistrationGUID(fmt.Sprintf(
+		"%v--%v--%v",
+		enclaveId,
+		serviceId,
+		time.Now().Unix(),
+	))
+	registration := user_service_registration.NewUserServiceRegistration(guid, enclaveId, serviceId, ipAddr)
+
+	backend.serviceRegistrations[guid] = registration
+	shouldRemoveRegistrationFromStore := true
+	defer func() {
+		if shouldRemoveRegistrationFromStore {
+			delete(backend.serviceRegistrations, guid)
+		}
+	}()
 
 	registrationsForEnclave[serviceId] = registration
-	shouldDeleteRegistrationEntry := true
-	func () {
-		if shouldDeleteRegistrationEntry {
+	shouldRemoveRegistrationFromIndex := true
+	defer func() {
+		if shouldRemoveRegistrationFromIndex {
 			delete(registrationsForEnclave, serviceId)
 		}
 	}()
 
 	shouldFreeIp = false
-	shouldDeleteRegistrationEntry = false
+	shouldRemoveRegistrationFromStore = false
+	shouldRemoveRegistrationFromIndex = false
 	return registration, nil
 }
 
-func (backend *DockerKurtosisBackend) GetUserServiceRegistrations(ctx context.Context, filters *user_service_registration.UserServiceRegistrationFilters) (map[user_service_registration.ServiceID]*user_service_registration.UserServiceRegistration, error) {
-	//TODO implement me
-	panic("implement me")
+func (backend *DockerKurtosisBackend) GetUserServiceRegistrations(
+	ctx context.Context,
+	filters *user_service_registration.UserServiceRegistrationFilters,
+) (map[user_service_registration.UserServiceRegistrationGUID]*user_service_registration.UserServiceRegistration, error) {
+	backend.serviceRegistrationMutex.Lock()
+	defer backend.serviceRegistrationMutex.Unlock()
+
+	result := map[user_service_registration.UserServiceRegistrationGUID]*user_service_registration.UserServiceRegistration{}
+	for enclaveId, serviceRegistrationsForEnclave := range backend.serviceRegistrationsIndex {
+		if filters.EnclaveIDs != nil && len(filters.EnclaveIDs) > 0 {
+			if _, found := filters.EnclaveIDs[enclaveId]; !found {
+				continue
+			}
+		}
+
+		for serviceId, serviceRegistration := range serviceRegistrationsForEnclave {
+			if filters.ServiceIDs != nil && len(filters.ServiceIDs) > 0 {
+				if _, found := filters.ServiceIDs[serviceId]; !found {
+					continue
+				}
+			}
+
+			registrationGuid := serviceRegistration.GetGUID()
+			if filters.GUIDs != nil && len(filters.GUIDs) > 0 {
+				if _, found := filters.GUIDs[registrationGuid]; !found {
+					continue
+				}
+			}
+
+			result[registrationGuid] = serviceRegistration
+		}
+	}
+	return result, nil
 }
 
 func (backend *DockerKurtosisBackend) DestroyUserServiceRegistration(ctx context.Context, filters *user_service_registration.UserServiceRegistrationFilters) (resultSuccessfulServiceIds map[user_service_registration.ServiceID]bool, resultErroredServiceIds map[user_service_registration.ServiceID]error, resultErr error) {