+++ container-engine-lib/lib/backend_impls/docker/object_attributes_provider/label_key_consts/label_key_consts.go
@@ -29,8 +29,10 @@ const (
 
 	isNetworkPartitioningEnabledKeyStr = labelNamespaceStr + "is-network-partitioning-enabled"
 
-	// TODO Remove this and instead use container domain names so we're not dependent on static IPs!
 	privateIpAddrLabelKeyStr = labelNamespaceStr + "private-ip"
+
+	// The GUID of the user service registration that a user service consumes
+	userServiceRegistrationGuidLabelKeyStr = labelNamespaceStr + "service-registration-guid"
 )
 // !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! DO NOT CHANGE THESE VALUES !!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 // If these value change, it will lead to the Kurtosis engine losing track of old containers
@@ -49,3 +51,4 @@ var PortSpecsLabelKey = docker_label_key.MustCreateNewDockerLabelKey(portSpecsLa
 var EnclaveIDLabelKey = docker_label_key.MustCreateNewDockerLabelKey(enclaveIdLabelKeyStr)
 var IsNetworkPartitioningEnabledLabelKey = docker_label_key.MustCreateNewDockerLabelKey(isNetworkPartitioningEnabledKeyStr)
 var PrivateIPLabelKey = docker_label_key.MustCreateNewDockerLabelKey(privateIpAddrLabelKeyStr)
+var UserServiceRegistrationGUIDLabelKey = docker_label_key.MustCreateNewDockerLabelKey(userServiceRegistrationGuidLabelKeyStr)